% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallmarks_V7.2}
\alias{hallmarks_V7.2}
\title{Hallmarks version 7.2 gene list set}
\format{
An object of class \code{list} of length 50.
}
\usage{
hallmarks_V7.2
}
\description{
Hallmarks version 7.2 gene list set
}
\keyword{datasets}
