% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huva.db}
\alias{huva.db}
\title{Collection of human expression profiles}
\format{
An object of class \code{huva_dataset} of length 4.
}
\usage{
huva.db
}
\description{
Dataset of expression matrix and paired annotation and metadata of several independent studies including both PBMC and sorted immune cell populations
}
\details{
Included datasets:
500FG - PBMC
ImmVar - CD4 T cells and CD14 Monocytes
CEDAR - Several
Classifier - PBMC
}
\keyword{datasets}
