% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuVa_db_overview.R
\name{huva_overview}
\alias{huva_overview}
\title{Getting the overview table of the dataset}
\usage{
huva_overview()
}
\value{
Overview table
}
\description{
Getting the overview table of the dataset
}
\examples{
huva_overview()

}
