% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_V7.2}
\alias{msigdb_V7.2}
\title{MsigDB version 7.2 gene list set}
\format{
An object of class \code{list} of length 25724.
}
\usage{
msigdb_V7.2
}
\description{
MsigDB version 7.2 gene list set
}
\keyword{datasets}
