% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overview}
\alias{overview}
\title{Overview table of the available datasets}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
overview
}
\description{
Overview table of the available datasets
}
\keyword{datasets}
